fix: move the endpoint guard to aiToolkit's public root so services stop deep-importing internal/
`evaluateSecretEndpoint`/`assertSecretEndpoint` — the SSRF/key-exfiltration
guard that decides whether a paid API key may be attached to a provider
`endpoint` — lived at `aiToolkit/internal/endpointGuard.js`, yet PortOS
service modules imported it from that private path. Any toolkit resync or
internal reshuffle would have required a manual multi-service migration,
with a runtime import failure as the failure mode.

Move it to `aiToolkit/endpointGuard.js`, a public peer of
`errorDetection.js` — the module that already establishes the "pure toolkit
root module, direct-importable and barrel re-exported" shape. `index.js`
adds the matching `export *`, so the barrel advertises the surface without
any consumer paying for the composition root's route/provider/runner graph
to reach one pure URL check.

Guard logic and its test are byte-identical to before: same allow/block
decisions, same evaluate-vs-assert split, same keyless local-LLM behavior,
same `allowCustomEndpoint` opt-in.

Fixes #5619
